TIP-607: add a proxy normalizer for Group and Variant
