+++ src/Pim/Bundle/CatalogBundle/Resources/config/serializers_standard.yml
@@ -16,6 +16,7 @@ parameters:
     pim_catalog.normalizer.standard.channel.class: Pim\Component\Catalog\Normalizer\Standard\ChannelNormalizer
     pim_catalog.normalizer.standard.currency.class: Pim\Component\Catalog\Normalizer\Standard\CurrencyNormalizer
     pim_catalog.normalizer.standard.family.class: Pim\Component\Catalog\Normalizer\Standard\FamilyNormalizer
+    pim_catalog.normalizer.standard.proxy_group.class: Pim\Component\Catalog\Normalizer\Standard\ProxyGroupNormalizer
     pim_catalog.normalizer.standard.group.class: Pim\Component\Catalog\Normalizer\Standard\GroupNormalizer
     pim_catalog.normalizer.standard.group_type.class: Pim\Component\Catalog\Normalizer\Standard\GroupTypeNormalizer
     pim_catalog.normalizer.standard.locale.class: Pim\Component\Catalog\Normalizer\Standard\LocaleNormalizer
@@ -128,12 +129,18 @@ services:
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
 
+    pim_catalog.normalizer.standard.proxy_group:
+        class: '%pim_catalog.normalizer.standard.proxy_group.class%'
+        arguments:
+            - '@pim_catalog.normalizer.standard.group'
+            - '@pim_catalog.normalizer.standard.variant_group'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
     pim_catalog.normalizer.standard.group:
         class: '%pim_catalog.normalizer.standard.group.class%'
         arguments:
             - '@pim_catalog.normalizer.standard.translation'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
 
     pim_catalog.normalizer.standard.group_type:
         class: '%pim_catalog.normalizer.standard.group_type.class%'
@@ -156,5 +163,3 @@ services:
         class: '%pim_catalog.normalizer.standard.variant_group.class%'
         arguments:
             - '@pim_catalog.normalizer.standard.translation'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
